+++ changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * New borrowing request.
     * End borrowing request.
     * Add inventory item request.
+    * Patch inventory item request.
 
 ### Removed